Cleaned up exits and movement handling
We now have a map for storing exits from a location and generic move and set
exit routines. Exits could be an array with constants like NORTH for the
indexes but at the moment we can have any exit for example: GO NORTH, GO DOOR,
GO STAIRS. However we could also have the exits as an array for N, NE, E, SE,
S, SW, W, NW, N, UP, DOWN and have any others as special cases. In the Java
version we have narrativeDoor, narrativeDoorLockable and narrativeNamedExit so
maybe the array route it will be?

Signed-off-by: Andrew Rolfe <andrew@bulldogdigital.co.uk>