Switch exits to an array
basic.exits is now an array instead of a map. The array has 10 elements for
the directions N, NE, E, SE, S, SW, W, NW, UP and DOWN. To make access easy
there are long and short constants defined for the indexes. So you can
reference an exit in a specific direction using either of:

  exits[N]
  exits[NORTH]

To go the other way - get a direction from an index - there is the exitNames
array. So either of these will return the string "North":

  exitNames[N]
  exitNames[NORTH]

Updated Move, Exits and SetExit functions to use new exit type and lookups.
Updated bootstrapping server to use new facilities as well.

Signed-off-by: Andrew Rolfe <andrew@bulldogdigital.co.uk>