player: Avoid double locking on new players
Currently we lock the location to add the player into the world and then lock
the location again to perform a 'LOOK'. By locking the location once - which
we currently do - and them manually handling the 'LOOK' command the command
can rely on our initial location lock. This helps lower contention for
location locks if there are many players trying to connect simultaneously.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>