world,loader: Drop location references from world
Currently the world package is being imported iinto the loader so that all of
the loaded locations can be added to the world by calling world.AddLocation().
However nothing is ever done with the world's locations. Locations can be
found and navigated using the starting locations and using the interconnecting
exits as a wierd and wonderful multiply linked list.

So we can simplyfy things by dropping world.AddLocation() and world no longer
needs to import the location package. The loader no longer need to import
world to call world.AddLocation(). So delete, delete, delete :)

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>