world,server: Tidy up world creation
Soft coded HOST and PORT in world and we now pass them in as parameters from
the server instead of hard coding them as constants - they will be
configurable in the server soon anyway.

Tidy up the logging so world messages reference the world and server messages
reference the server.

Dropped TODO comment from server as multiple worlds will not be supported.
It's more sensible to run multiple servers on different ports so if one goes
down all the servers/worlds don't go down.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>