inventory: 'shrink' inventories that become empty
When things are put into an inventory the length and capacity of the slice
used to hold the content is increased. However when things are removed the length is
reduced but the capacity is not. As locations use inventories this could
potentially waste a lot of storage - and appear as a memory leak. This fix
zeros out the content slice when an inventory's last item is removed.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>