client,sender: Add prompt control
This commit adds Prompt to the sender interface and implements it in client.
Prompt will allow a new prompt to be set and return the old prompt. In effect
allow a prompt to be switched to something else and then back again.

The client uses this to set an empty prompt when quitting to avoid displaying
misleading/inappropriate prompts.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>