player: Fix 'memory leak' in playerList
When players are removed from the playerList the players slice length is
reduced but it's capacity is not reduced. This could be perceved as a memory
leak. When the last player disconnects we therefore reset the capacity of the
slice.

Also dropped pointless not found logging. It makes no difference if the player
is not in the list.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>