thing,uid: Separate UID from Thing
At the moment UID is very strongly tied to Thing. This makes it difficult to
access the UID in some instances as we need to import the thing package
causing import loops.

This change moves the methods IsAlso() and UniqueId() from Thing to UID. UID
now has an Interface that requires these methods and implements them for UID.
The uid.Interface is now embedded into thing.Interface and UID is embedded
into Thing. We can now examine and compare UID's just by knowing something
implements the uid.Interface without having to import the thing package.

Updated the thing and uid tests to move testing of IsAlso() and UniqueId()
from thing_test to uid_test

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>