client: Drop KLUDGE, fix receive() multi-commands
When there is a network delay two commands may be sent together. For example
we may receive "S\r\nS\r\n" together. At the moment we only expect one command
at a time and when process the example we don't understand it's two commands
run together so we just reply "Eh?".

This update fixes the above situation and as a result also drops the KLUDGE
required for Windows TELNET. We still don't put Windows TELNET into line mode
but we can handle character at a time in the mainline code. However there is
still the associated performance penalty.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>