client: Make MAX_TIMEOUT more flexible
Currently MAX_TIMEOUT is set to a number of minutes and then multiplied by
time.Minute buried in the code. This change moves the multiplication to the
definition of the MAX_TIMEOUT constant and is easily configured:

  MAX_TIMEOUT = 90 * time.Seconds
  MAX_TIMEOUT = 10 * time.Minute
  MAX_TIMEOUT = 1 * time.Hour

This also removes the multiplication from within the for loop which is always
a good idea - although being constants this may be optimised by the compiler.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>