client: Reduce Send() processing
Instead of always concatenating line feeds onto the format string put the
linefeed in the prompt constants and only do it once. Also instead of adding
the prompt as a parameter and concatenating a place holder for it onto the
data we can just concatenate the prompt onto the data directly. This avoids an
append and a replacement in the Sprintf() call.

NOTE: One downside of this is that just hitting return did return:

  >
  >

With these changes we now get:

  >

  >

Is this an issue?

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>