location,basic: Add notion of a crowd
Added a Crowded() method to Basic locations which returns a boolean depending
on whether a location is considered crowded or not. Currently a crowd is
determined by comparing the number of things in a location to CROWD_SIZE. This
method has also been added as part of location.Interface.

NOTE: Currently Crowded simply counts the number of things in a location - it
should probably only count mobiles.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>