basic: Reduce messages in crowded locations
If a location is crowded you are not likely to notice someone arrive or leave.
By not sending messages in these situations we can vastly reduce the amount of
network traffic generated on a busy server.

Also if a location is crowded you are not likely to take notice of everybody
that is there. So we simple display 'You can see a crowd here.' instead of
listing everyone.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>