player: Don't announce arrival in crowded locations
If a new player is spawned in a crowded location don't announce player's
arrival as they are likely to go unnoticed. This can drastically reduce
network traffic. For n players arriving one after the other and assuming none
leave in the meantime the number of messages needed to be sent is given by:

  ((n-1)n / 2)

So for 5,10,20 and 50 players we get:

   5: (4*5)   / 2 =    10
  10: (9*10)  / 2 =    45
  20: (19*20) / 2 =   190
  50: (49*50) / 2 = 1,225

We would also need to send this many messages again as each player leaves the
location. Currently for testing 5,120 players with only three starting
locations we send:

  5,120 / 3 = 1,706 players per start location

  (1,705*1,706) / 2 = 1,454,365 messages per location

  1,454,365 * 3 = 4,363,095 arrival messages in total

With a crowd size of 10 we reduce this from 4,363,095 messages down to 135
messages (45 messages * 3 locations). Each message is roughly 75 bytes so we
go from sending 312Mb of data down to just under 10Kb.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>