location,basic: Cater for no obvious exits
At the moment if there are no exits from a location the standard message is
displayed with no exit names appended to the end:

  You can see exits:

This change will display a more meaningful message:

  You can see no immediate exits from here.

So there may be hidden or non-obvious exits, or you could simply be trapped ;)

NOTE: We only keep a reference to the starting locations. These starting
locations have references to other locations - via exit pointers - which
reference further locations in the same way. This creates a chain of
locations. If a location or a chain of locations is not connected to a
starting location they will be unreferenced and garbage collected! This will
need to be addressed at some point.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>