basic: Fix newline/colors describing locations
When describing locations we always send:

  [CYAN]text[WHITE]⏎
  text⏎
  [GREEN](text)⏎  () = conditional
  ⏎
  text⏎

This always sets a green color before any objects are listed if there are any.
If there are no objects we still set the color to green and send a newline
before and after the color code resulting in double spacing. Instead this
should be:

  [CYAN]text⏎
  [WHITE]text⏎
  ([GREEN]text⏎)  () = conditional
  ⏎
  text⏎

Now we only send the green when it will be used along with the extra newline.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>