thing: Implement recordjar.Unmarshal & Init
Implemented recordjar.Unmarshal and added an Init function. This is so we can
unmarshal things from a recordjar which creates instances. Once we have all
instances available we can initialize them. This needs to be done in two
phases: Unmarshal/Init so that instances can reference each other - for
example when linking up location exits.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>