Drop unused New functions.
Now we are unmarshalling and initializing types it makes the New functions
redundant - so we can drop them. Any setup has been moved to the relevant Init
methods already. Now to create instances we can either use Unmarshal with a
Record or just create an instance of the appropriate struct directly. Then
call Init() for final setup.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>