recordjar: Fix explicit break/continue
When cleaning up the recordjar code bare break/continue statements had labels
added to them for readability as we already had the labels anyway. However due
to the placement of the labels the logic of the 'jumps' changed. Where we were
breaking out of the current loop using a bare break changing it to break
RECORDS was incorrect and should actually have been changed to continue
RECORDS to preserve the previous logic.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>