recordjar: Fix line ending encoding issues
This commit fixes issues with different line endings which prevented the
record separator from being detected on Windows platforms due to the specific
"%%\n" test.

This fix makes the encoding of line endings in *.wrj files irrelevant.

It also makes it so the only encoding we have to deal with is "\n". This
should mean we can simplify text.Fold(). We should also check for other
instances where we check for different line endings and change them to just
handle "\n" from now on.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>