command: Fix *Command.New("")
The whitespace fix in commit 694423 breaks *Command.New - as
opposed to the package level command.New - if the input string is empty. This
is because Split will return a string slice with one empty string element
while Fields will return an empty slice. In other words:

  fmt.Printf("Split: %q\n", strings.Split("", ` `))
  fmt.Printf("Fields: %q\n", strings.Fields(""))

Will produce:

  Split: [""]
  Fields: []

Another side effect is that we have to take more care resetting the Command
struct as we cannot rely on words[0] being set to an empty string and
propagated to other variables - we have to specifically reset them.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>