client: Simplified & improved bail locking
Instead of using bail and mutex variables we can just define bail as a
buffered channel of type error. Then we pre-load the channel with a nil error.
The bail channel now acts like the old mutex - read from the channel to take
ownership (lock) then release it by putting back into the channel (unlock).

This simplification allows us to remove the lock and unlock methods which are
now unnecessary. A bailed method has been added so that we can check for the
error that caused us to bail.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>