client: Make whitespace trimming a little more efficient
When extracting the command to be processed next from the buffer we only need
to trim whitespace from the end. This is more efficient than trimming both
ends and we can even call TrimRightFunc() directly. If we have whitespace only
input we get the same empty string result as trimming both ends would have
given us.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>