client: Improve buffer handling + fix runaway buffer size
Current processing with the input buffers causes a lot of allocations and also
does not place limits on how much data can be received - or how large the
input buffers can grow to. This can be exploited by a client sending large
amounts of data that do not contain a linefeed.

This change limits the input buffer - currently 256 bytes. It also improves
the buffer handling reducing allocations to a minimum. See copious comments in
the code for more information.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>