client: Improve scanning of buffer avoiding garbage
Having the LF < bCursor check in the for loop introduces an important check
that could be overlooked during future code changes.

There is an alternative to using LF < bCursor in the for loop. We can simply
use buffer[0:bCursor] in the nextLF function. This reduces the amount of data
the nextLF function has to 'scan' as well as cleaning up the for loop. This
also fits with using the nextLF function in the first place - to simplify the
for loop.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>