location,playerList: Fix data races
It seems that I did my locking the wrong way around. Data synchronisation
happens on a receive from a channel. So for our mutex we want the data
synchronised before we play with it. Therefore we need to swap the logic in
lock and unlock and pre-seed a value into the channel.

Found by Go 1.1 race detector.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>