command: Store original user input
For commands like 'say' and 'talk' we want to be able to use the user's
original input - not the all uppercase parsing version. This change adds an
input member to the Command struct. It is a []string and holds the input as
whitespace separated words. Added check for original input to tests.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>