client: Cleanup error and timeout checking
This change is an effort to clean up and simplify error and timeout checking.
Timeouts throw a spanner in the works as they are presented as errors - we
want to know about them but we don't want them handled as actual errors.

A hopeful side effect is more robust error handling that is easier to
maintain.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>