parser: Added Next method to interface
This change adds a Next method to the parser interface so that one parser can
hand over control to another parser. For example we can start with a login
parser and when we have a successful login we can transfer control to the main
player parser. This allows us to keep the different parsers modular and we can
easily add or replace parsers.

This change also adds a stub Next method to player so that it continues to
implement the parser interface. Also it's now a bad idea to set the parser's
sender to nil in the Destroy method as we will need to pass it to the next
parser.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>