client: Switch to parser.Next() processing
This change wraps the main receiver() processing in a parser.Next() loop. This
will allow different parsers to be swapped in and out.

The greeting should be displayed by a parser when initially connecting to the
server. So we drop it from the client package. We also drop explicit prompt
setting as we expect the parsers to know what prompts they want to set.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>