data: Added players/.gitignore
For now we create the players directory and add it to Git with a .gitignore
file to exclude the player data files from Git. This is being lazy for now.
What we should do is check for the data/players directory in code and create
it if it's missing. This would also let us change the location of the players
directory through the configuration settings.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>