player: Initial login parser
This parser currently provides a mechanism to allow players to login to the
server. Currently you can ONLY login - player accounts have to be created
manually in the data/players directory. For a discussion of manually creating
player files see:

  http://wolfmud.dev/journal/2013/7/22.html

Work is now ongoing to allow player accounts to be created when connecting to
the server.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>