player: Integrate Player parser with Login parser
Now we have the Login parser and can load player files we can drop the dynamic
player creation code. This created 'Player x' players each time someone
connected to the server. So now we can drop the playerCount package variable,
then init function and the loadPlayer function. A players data is then passed
to the player parser as a RecordJar from the Login parser and we just have to
unmarshal it.

We also update the Next method to return to the Login parser when the player
quits.

NOTE: At the moment we do not check to see if the player is already logged in.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>