loader: Added Unmarshaler registration
Unmarshaler registration is going to be used to register recordjar unmarshalers
with a string key. For example a thing may register using:

        loader.Register("thing", &Thing{})

This registers a zero value Thing struct - which must implement the Unmarshaler
interface - with the key "thing". When reading recordjar files each record
should have a type attribute. The type attribute is then looked up against the
registered keys to find the record's unmarshaller.

For now this commit just adds the registration functionality.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>