is: New package for interfaces
The 'is' package is going to be used to generalise the entities and other
types a lot more - more Go style and less OOP style. At the moment we just
have the Nameable interface. Due to package naming we can now do:

        if n, ok := x.(is.Nameable); ok {
                fmt.Printf("You see %s", n.Name())
        } else {
                fmt.Printf("You see something...")
        }

The 'is.Nameable' reads quite nicely and I didn't know what else to call the
package otherwise. There is a Go convention to use an -er suffix for single
method interfaces. However for the types of interfaces needed, what they
represent and how they interact with entities an -able or -ible suffix is
going to make more sense: is.Nameable, is.Describable, is.Edible,
is.Examinable etc...

common.go is going to hold general interfaces anything might implement. There
will soon be a command.go also for command interfaces. For example:

        is.Examinable
        is.Gettable
        is.Dropable
        is.Weighable

For some insight on my thinking the following journal entries may shed
some illumination:

        http://wolfmud.dev/journal/2013/8/22.html
        http://wolfmud.dev/journal/2013/8/28.html

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>