recordjar: Move Init from thing to Unmarshaler interface
The Init function should not be part of the thing.Interface interface, it
should be part of the Unmarshaler interface. This is because unmarshalers
should perform two functions: Unmarshal - Take recordjar records and setup
associated type structure. Init - perform any initialisation and setup after
all records for a recordjar have been unmarshaled. Splitting unmarshaling into
two functions allows all of the types to be created and then during
initialisation and setup they can refer to each other. For example you need to
load all of the locations and then link up all the exits - which requires
referencing other locations. Another example is creating items and then
putting them into an inventory - the inventory must exist to put the items
into.

The Init function should also deal with a map of recordjar.Unmarshaler instead
of thing.Interface to keep things general. This means we now need a map of
recordjar.Unmarshaler instead of thing.Interface for the refs. We also need to
use the is.Nameable assertion on the refs map elements to see if we can get a
name for an element - previously this was available from the thing.Interface.

So this commit moves the Init function to the recordjar.Unmarshaler interface
and updates it's signature. It also updates all of the currently implemented
Init functions to use the new map type. Finally loader.load is updated to use
is.Nameable() on the refs map elements.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>