loader: Switch to Unmarshaler general handling & loading
First thing this commit does is add init functions to all of the entities so
that they register themselves with the loader. This requires a 'side effect'
import in the server package to trigger the item init as nothing directly
refers to it.

Second thing this commit does is rework the loader so that entities are
dynamically loaded using the registered types - goodbye type switch :)

NOTE: This commit does not affect player loading yet which is still a simple
recordjar unmarshaling.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>