player: Change internal players slice to a map
This commit changes the internal players list representation from a slice of
player pointers to a map of player pointers. The intent is to have the account
id as the map key - at the moment it's the player's name until we can create
accounts properly. Having the PlayerList keyed lets us efficiently check for
duplicate logins when adding players to the list - for which we now return a
DuplicateLoginError, although we are not checking for it just yet.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>