player: Added main driver + game & menu drivers
To avoid confusion between frontend parsers and ingame parser the frontend
parsers are now called drivers. Currently we have a game driver and a menu
driver. Drivers are based on the main driver so that adding a new driver is
simplified.

Note that this commit just adds the new drivers, we have not switched over to
using them yet although login contains a small newLogin stub for compiling the
current code.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>