client,player: Reimplement login as a driver
This commit reimplements the login parser as a login driver. We now have
login, menu and game drivers and can switch over to actually using them.

This commit also reworks client to handle a driver instead of a parser. It
also drops the Next method and parser import from player - the game driver now
talks to player and player no longer needs to implement a parser.

                             + + + WARNING! + + +

At the moment there is an issue of a player getting 'stuck' if they login to
the menu and then the connection is dropped. They player ends up in limbo and
further logins result in duplicate logins being detected. The only solution
currently is to restart the server. This is obviously unacceptable and is
currently being working worked on.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>