command: Prevent message colours bleeding into next message
When messages are buffered it is possible for the color set in one message to
bleed into the next message if: 1st message does not reset the colour or 2nd
message assumes default colour is set.

This commit adds a "[WHITE]" reset to the end of messages automatically to
reset any other colours set. This also means we don't have to have "[WHITE]"
littering the code to reset the colours manually.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>