client: Simplify network error checking
A timeout is not actually an error and usually indicates an inactive player
session due to our setting of SetReadDeadline. We can therefore simplify our
error handling by not treating a timeout as a special kind of error with it's
own logic. Instead we accept it is a non-error, drop the special handling and
simplify things.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>