player: Roll Destroy and remove methods into quit
There is no need to separate quit and Destroy processing any more. So we roll
the Destroy - and remove - method functionality into the quit method. We also
need to update game.go to drop the Destroy call and client.go to use the
Logout instead of Process method.

Now we log off the account calling Logout and logout the player from the game
with quit.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>