driver,player: Start using account not name
It is about time we started using 'account' to refer to the player's data file
name and their login credentials instead of using 'name'. 'name' should be
reserved for the player's in game character name and 'account' for the
player's data file name only.

Why the distinction?

Accounts have to be unique, character names do not.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>