driver,player: Store account as SHA512 hash
Instead of storing the account as a string store it as a SHA512 hash. We need
to compute the hash anyway to compare with the player's data file name. This
also means we are not storing arbitrarily long strings passed to us by clients
which could be used to attack the server by trying to use excessive amounts of
memory.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>