recordjar: Added Decoder type + decoder refactoring
Currently all of the methods in decoder.go take a Record as a receiver. This
causes problems when we want to add an Encoder such that the methods are named
the same in both the encoder and decoder so they compliment each other well.

This change adds a Decoder type that has the same structure as Record. This
allows for easy conversions from Record to Decoder and back. All of the
current methods in decoder.go now take a Decoder where they previously took a
Record. The parameters for Unmarshal and Init methods in recordjar.go have
also been updated to user the new Decoder type. All references to and usage of
Unmarshal and Init have been updated as well.

Now we can now define methods on the Decoder - and soon the Encoder - type
without naming issues. Methods like DecodeKeyword and EncodeKeyword were
really not going to fly...

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>