utils: fix max concurrent players stats
When stats are updated the current number of players is compared to the last
maximum number of concurrent players. If players login and quit between an
update of the stats the maximum concurrent players will be incorrect because
the stats don't 'see' the player who nipped in and out again between updates.

To fix this the maximum concurrent players is now checked and recorded by the
player list when the player logs in. When the stats are updated the player
list is queried for the player stats.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>