driver,player: Refactor driver-player interaction
This commit is quite large but simplifies how the drivers interact with a
player instance. Currently we accumulate player data and store it in a
driver instance along with a pointer to a player instance. This is pointless
and we now just store the values directly into the player instance.

This also means we need to move the guts of account.create() into
player.save() - really it should be moved into a proper Marshler method but we
are not ready for that just yet.

All of the player functions are now methods on a player type.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>