entities,driver: Move validation to player from account
This change moves the validation done by the account driver into the player
type. For example it's more logical to have the validation for a password in
the SetPassword method - this ensures the validation is done every time the
password is changed. The same goes for validating account IDs.

The only reason this wasn't done initially was because all of the validation
was written when the account driver was written and so all of the focus was on
the driver code at the time.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>