attr: Have writing implement description i'face
Make writing implement the description interface. Then commands like 'EXAMINE'
can append the additional description about the writing instead of having in
the actual description. This allows us to have e.g. a mug:

  This is a large, white, chipped mug.

Then we can write on it and the writing attribute will add to the description:

  This is a large, white, chipped mug. It has something written on it.

Without this we would need some way of pulling apart the main description if
we wanted to manipulate it.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>