attr: Drop reflect and thing.Find
Originally thing.Find was introduced to cut down the repeated coding required
for the attribute finders. It worked well and was a nice idea. However using
reflection caused the performance of every finder to drop by around 75%. This
change drops thing.Find and reflection and goes back to using simple
assertions instead.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>