attr: Switch to nil pointer types
Warning! This is a bit of an experiment :)

There are times when you want a method on a type instead of a function.
Examples would be NewXXX and FindXXX functions:

  NewAlias() *alias
  FindAlias(has.Thing) []has.Alias

This creates a lot of package level functions. However it would be nice if we
had New and Find methods on the types without an actual instance of the type.
We can call methods on nil pointers. So for each attr type a nil pointer of
the correct type can be created very cheaply (memory and performance) and can
be used to call the methods:

  func Alias() *alias { return nil }
  func (*alias) New() *alias
  func (*alias) Find() []has.Alias

Note that the receiver is unnamed to indicate it is unused (it's nil!). Also
the nil pointer is created using a function as the type is not exported. Now
we can do things like:

  Alias().New("test")
  Alias().Find(t)
  Name().New("test")
  Name().Find(t)

It also means that we could/might include the methods in interfaces which we
couldn't before. Not sure if this will pan out but it seems like a resonable
idea at the moment.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>