cmd: Add 'comma ok' idiom to command processing
Added an ok return variable to commands so that they can indicate if the
command was successful or not. For example in move.go we now have:

  if msg, ok = exits.Move(t, cmd); !ok {
    return
  }

This also makes it easier for commands to call each other to do some
processing instead of duplicating code. For example when executing a PUT we
can issue a DROP before moving the item to a container which takes care of
checking if the item can actually be dropped and isn't vetoed for some reason.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>