cmd: Drop unused utils.what function
This function didn't really work as well as intended. Frequently access was
needed to the inventories after calling what which meant a second attribute
lookup to find it in the calling code. This affected performance quite a bit.
The equivalent of what has now been inlined in the relevant commands. This has
two benefits: Inventories looked up can be reused and the sequence that
inventories and narratives are visited can be varied depending on the command
and it's requirements.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>