attr: Swap inventory List and Contents method names
There is already an Exits List method used to list out the exits for a
location. For consistency it makes sense that List on an Inventory lists it's
content. This change swaps the method names for List and Contents over. It
also makes more sense for Contents to return the Inventory actual content
objects (method now matches the struct field name).

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>