attr: Update Exits to use Inventory references
Exits and locations are all about Inventory attributes. To be at a location
you are in that location's inventory. To move you are removed from the current
inventory and added to the destination inventory. It therefore makes more
sense to link Exits via Inventory instead of Thing references.

This has allowed us to drop a relitivly expensive FindInventory call in the
commonly used Move method at the expense of additional FindInventory and
Parent calls in code mostly used during setup and initialisation.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>