attr,cmd: Locate now uses an Inventory reference
This changes the Locate attribute to use the Inventory interface instead of a
Thing interface. The Locate interface methods Where/SetWhere have been updated
to take/return Inventories to match.

The command changes are due to the has.Locate interface changing. The commands
all work and compile but need reviewing and cleaning up at this point.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>