server: Simplify input handling
This change switches from using a buffio.NewReader to using a bufio.NewScanner
to handle player input. This also means that '\n' is not appended to the end
of the string passed to cmd.Parse any more.

There is a minor, temporary bug caused by this change: exiting the game by
typing 'quit' prints 'Eh?' before exiting.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>