cmd: Switch parse to use new state
Switch parser to use the new state type. Note that commands have not been
updated to use the new state type yet. To work around this in the dispatch
function we directly interact with the state to extract the data commands are
expecting and to push data back into the state from the commands. This is only
temporary and as the commands are updated the dispatch function will be
cleaned up.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>