cmd: Add new quit command
This change adds a new quit command. Before 'quit' just flagged the server
input processing loop to exit. This caused 'Eh?' to be displayed before the
server quits due to the 'quit' command not being understood.

Now we have a proper quit command handler. Currently it just displays the
message 'Bye bye...'. In future this handler will provide a hook into the quit
process to do things like veto quitting if a player is engaged in combat.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>