attr: Add NormalizeDirection method to exits attribute
This change adds the NormalizeDirection method. This method takes a direction
and returns the direction's long name in lower case. For example: 'N',
'NORTH', 'n', 'north', 'North' and 'NoRtH' will all return 'north'. This
method uses the directionIndex map which has been expanded with all lower,
upper and title cases for directions. Using a quick map look of the string is
preferable but strings.ToUpper will be called as a fallback e.g. in the case
of 'NoRtH'.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>