cmd: Reimplement move command to use parser state
This change updates the move command handler to use the new parser state. It
also merges the exits.Move method into the move handler. This simplified the
update to use the parser state and moves all of the move code into the
handler - which is where it should have been put.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>