Reduce string <-> []byte conversions for out bound text
Try and use []byte as much as possible. A lot of the time we are
manipulating/concatenating strings and causing unnecessary allocations,
strings being immutable. The change to parse and fold only have an effect on
outgoing text. This will be important for the networking changes as we will
want to write bytes, not strings converted to bytes.

Ironically we introduce an allocation in Fold using bytes.Runes. This should
be offset by replacing 'in += "\n"' with "in = append(in, '\n')". Did the
range over the string automatically incur an allocation?

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>