cmd: Split state's original input into words as well
This change splits the original input into words the same way we handle
splitting the input into uppercased words. This means there will be an easier
1:1 mapping between the original input []string and the uppercased words
[]string.

This is more for commands such as SAY (not implemented yet) where we want
access to the original input. As an example "say diddymus Hello, World!", we
want to uppercase this for processing the SAY and DIDDYMUS but we want the
original message to stay as 'Hello, World!' and not 'HELLO, WORLD!'.

Alternatively we could keep just the original input split into words and only
uppercase the slice elements as and when we need them for processing the command,
saving some CPU cycles on unprocessed words. Maybe something to look into once
everything is actually working?

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>