attr: Fix racy brl
The current BRL had a mutex for locking and a uint64 for the lock ID. We want
to be able to read/write the lock for a BRL but at the same time read the ID
with or without the lock being held. This is identified by the race detector
as being racy even though we know the ID will never be changed once the BRL is
created.

This update switches to two channels, one for locking and one for the lock ID.
These channels can operate independantly giving us the functionality required
without being racy.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>